Conflicts:
	set-definitions/src/main/java/eu/europeana/set/definitions/model/impl/BaseUserSet.java
	set-mongo/src/main/java/eu/europeana/set/web/model/WebUserSetImpl.java
	set-web/src/main/java/eu/europeana/set/web/service/UserSetService.java
	set-web/src/main/java/eu/europeana/set/web/service/controller/BaseRest.java
	set-web/src/main/java/eu/europeana/set/web/service/impl/UserSetServiceImpl.java

+++ set-mongo/src/main/java/eu/europeana/set/mongo/service/PersistentUserSetServiceImpl.java
@@ -191,7 +191,8 @@ public class PersistentUserSetServiceImpl extends AbstractNoSqlServiceImpl<Persi
 	@Override
 	public void remove(String id) {
 		PersistentUserSet userSet = getByIdentifier(id);
-		getDao().delete(userSet);
+		if (userSet != null)
+		    getDao().delete(userSet);
 	}
 
 	@Override
